pyproject.toml
[tool.poetry]
name = "home_recruiters"
version = "0.1.0"
description = "Every person needs help from other professionals be it electricians, plumbers, mechanics or any other. This project is a quick and easy way for people(recruiters) to connect with skilled people who think are a good fit and hire them"
authors = ["Leulseged Tariku <leulsegedtariku@gmail.com>", "Mohamed Nesredin <m.n.kaizen@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.2"
fastapi = {extras = ["all"], version = "^0.53.2"}
typer = "^0.1.1"
gunicorn = "^20.0.4"
pendulum = "^2.1.0"
confusable_homoglyphs = "^3.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
pyjwt = "^1.7.1"
loguru = "^0.4.1"
tortoise-orm = "^0.16.4"
python-dotenv = "^0.12.0"

[tool.poetry.dev-dependencies]
bandit = "^1.6.2"
black = "^19.10b0"
pytest = "^5.4.1"
coverage = "^5.0.4"
isort = "^4.3.21"
radon = "^4.1.0"
pytest-sugar = "^0.9.2"
pytest-cov = "^2.8.1"
safety = "^1.8.7"
mkdocs = "^1.1"
mkdocs-material = "^4.6.3"
mypy = "^0.770"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
